requirements.txt
httpx
pynacl